### 2026-04-07 #766 @IgnatiusPang
HydraR: Stateful Agentic Orchestration for Scientific Reproducibility in R
Submitting Author Name: Ignatius Pang
Submitting Author Github Handle: @IgnatiusPang
Other Package Authors Github handles: @aidantay
Repository:  https://github.com/APAF-bioinformatics/HydraR
Version submitted: [v0.2.4](https://github.com/APAF-bioinformatics/HydraR/releases/tag/v0.2.4)
Submission type: Standard

Version accepted: TBD

---

```
Package: HydraR
Title: Stateful Agentic Orchestration for Scientific Reproducibility
Version: v0.2.4
    c(person(given = c("Chi", "Nam", "Ignatius"),
             family = "Pang",
             email = "ignatius.pang@mq.edu.au",
      person(given = "Aidan",
             family = "Tay",
             comment = c(ORCID = "0000-0003-1315-4896")))
Description: A high-performance framework for orchestrating complex
    agentic workflows in R, specifically designed for scientific
    reproducibility and auditability. HydraR provides a robust,
    state-managed engine for Directed Acyclic Graphs (DAGs) and
    iterative state machines, prioritizing CLI-native LLM
    interactions (e.g., Gemini, Claude, Copilot) and hardened
    state persistence (DuckDB/SQLite). It features isolated
    execution via Git Worktrees for parallel file-modifying
    tasks, autonomous quality control through integrated
    auditors, and human-readable visualization using Mermaid.js.
    Ideal for creating auditable, resumable research assistants
    and complex bioinformatics pipelines.
License: LGPL (>= 3)
    DBI,
    digest,
    furrr,
    igraph,
    R6,
    yaml
    duckdb,
    httr2,
    future,
    devtools,
    withr,
    reticulate,
    DiagrammeR
```

    - [ ] rOpenSci internal tools

    HydraR provides a stateful, auditable engine for orchestrating complex agentic workflows, specifically targeting scientific reproducibility via Git worktree isolation and persistent DuckDB/SQLite checkpointing. It acts as a robust wrapper for large language model (LLM) command line interfaces (CLIs) tailored for bioinformatics research.

    The target audience includes bioinformaticians, research software engineers, and data scientists building multi-agent systems for automated data analysis, code refactoring, and large-scale literature synthesis.

While 'ellmer' focuses on high-level API convenience, 'mall' excels at structured data operations, and 'gptstudio' on IDE integration, HydraR is **uniquely centered on persistent session lifecycles** and **native command-line integration**. It enables complex research pipelines with **restorable execution states** and **complete audit trails**, permitting safe parallel file modifications via Git worktrees, which are features essential for reproducible scientific computing. Furthermore, unlike solutions that rely on 'reticulate' to bridge R with Python-based AI frameworks (such as LangChain), HydraR is **built entirely on R's native architectural frameworks**, avoiding the data-translation barriers and compatibility issues often encountered when passing complex statistical structures between languages.

    N/A

    N/A

    The package currently passes `R CMD check` with 0 Errors and 0 Warnings. Minor NOTEs regarding hidden developer files (`.github`, `.lintr`) are intentional for repository management.

## Use of Generative AI

- [x] Generative AI tools were used to produce some of the material in this submission.

If so, please describe usage, and include links to any relevant aspects of your repository.

The core R6 architecture, state-management patterns, and Git worktree isolation strategies were designed and authored by the human authors (Ignatius Pang and Aidan Tay) and then **rigorously tested manually**. AI (Antigravity and Gemini CLI) was strategically employed for implementing logic blocks, unit tests, and documentation, following a rigorous "Human-in-the-loop" pattern where every line was manually reviewed and verified. Detailed disclosure is available in **[agents.md](agents.md)** and **[DESIGN.md](DESIGN.md)**. To ensure software reliability, the framework is continuously evaluated against an automated testing suite consisting of 556 unit tests, currently maintaining a 100\% pass rate (zero failures).

*Note: We also intend to submit to JOSS as the primary scholarly venue for the software.*
