### 2026-04-07 224ec01
docs: Update README and JOSS submission with paper insights and rOpenSci standards
