+++ README.md
@@ -15,33 +15,54 @@
   </tr>
 </table>
 
-`HydraR` is a lightweight, state-of-the-art orchestrator for building general-purpose agentic workflows in R. It prioritizes **CLI-native LLM interactions**, **hardened state management**, and **graph-based execution** (supporting both Directed Acyclic Graphs and iterative loops).
+### *Stateful Agentic Orchestration for Scientific Reproducibility*
 
-## Why HydraR?
+`HydraR` is a lightweight orchestrator for building **autonomous AI workflows** natively in R. It allows researchers to design complex, multi-step processes—such as data cleaning, analysis, and report generation—that are driven by Large Language Models (LLMs) while maintaining the rigorous auditability and reproducibility expected in scientific research.
 
-Standard agentic frameworks often rely heavily on brittle API wrappers and volatile state. `HydraR` is built for durability and reproducibility:
-- **CLI-First**: Directly drive high-performance CLI tools like `gemini-cli`, `claude-code`, or `gh copilot`.
-- **Hardened State**: Implements a robust state machine with persistent checkpointing (DuckDB/SQLite).
-- **Graph-Native**: Design complex logic transitions and loops with built-in validation.
+![HydraR Visual Blueprint: A sample travel itinerary workflow showing iterative validation loops and parallel generation paths.](paper/figures/travel_workflow.png)
+
+## 🎯 What is HydraR? (For Non-Experts)
+
+Think of an **"Agent"** as an AI-powered assistant that can perform specific tasks. Instead of just "chatting" with an AI, `HydraR` allows you to link multiple assistants together into a **Workflow**. 
+
+You create a "blueprint" (a visual map) of your research tasks. `HydraR` then manages the execution: it gives each assistant its own isolated workspace, remembers every step they take (checkpointing), and allows them to "loop back" and fix mistakes if a task fails. This makes AI research **reliable, auditable, and automated**.
+
+## ⚖️ Why HydraR? (Statement of Need)
+
+Standard agentic frameworks often rely heavily on brittle API wrappers and volatile state. `HydraR` addresses three fundamental challenges in scientific AI orchestration:
+
+1.  **🔍 Auditability**: It maintains a central memory system that continuously saves workflow checkpoints (using a DuckDB database), ensuring a complete, verifiable record of all AI interactions and data changes.
+2.  **🛡️ Filesystem Safety**: By leveraging **Git worktrees**, `HydraR` creates temporary, isolated project branches for each agent working in parallel. This prevents file corruption and race conditions when multiple tasks run simultaneously.
+3.  **🏗️ Portable Workflows**: By defining agent logic using standard **YAML** files and **Mermaid.js** diagrams, `HydraR` separates the workflow's design from the code that runs it. This text-based blueprint acts as an "OpenAPI for Agents"—a shareable format that is easy to audit and version-control.
+
+## 📖 Glossary of Technical Terms
+
+To help researchers from different backgrounds, here are the key technical terms used in `HydraR`:
+
+- **Agent**: An autonomous AI entity (powered by an LLM) that can perform specific tasks or make decisions.
+- **Workflow**: A collection of tasks (nodes) linked together to achieve a larger goal.
+- **DAG (Directed Acyclic Graph)**: A mathematical structure used to represent workflows where tasks move in a specific direction without looping back infinitely.
+- **State Machine**: A system that keeps track of its current "status" and moves between states based on inputs or task results.
[…]
