### 2026-04-07 ef2c029
chore: version bump to 0.2.1 for submission release
