### 2026-04-07 cbea94b
chore: remove paper directory from branch main
