+++ paper/joss_submission/paper.tex
@@ -1,619 +1,0 @@
-\documentclass[10pt,a4paper,onecolumn]{article}
-\usepackage{marginnote}
-\usepackage{graphicx}
-\usepackage{xcolor}
-\usepackage{authblk,etoolbox}
-\usepackage{titlesec}
-\usepackage{calc}
-\usepackage{tikz}
-\usepackage{hyperref}
-\hypersetup{colorlinks,breaklinks,
-            urlcolor=[rgb]{0.0, 0.5, 1.0},
-            linkcolor=[rgb]{0.0, 0.5, 1.0}}
-\usepackage{caption}
-\usepackage{tcolorbox}
-\usepackage{amssymb,amsmath}
-\usepackage{ifxetex,ifluatex}
-\usepackage{seqsplit}
-\usepackage{fixltx2e} % provides \textsubscript
-\usepackage[
-  backend=biber,
-%  style=alphabetic,
-%  citestyle=numeric
-]{biblatex}
-\bibliography{paper.bib}
-
-
-
-% --- Page layout -------------------------------------------------------------
-\usepackage[top=3.5cm, bottom=3cm, right=1.5cm, left=1.0cm,
-            headheight=2.2cm, reversemp, includemp, marginparwidth=4.5cm]{geometry}
-
-% --- Default font ------------------------------------------------------------
-% \renewcommand\familydefault{\sfdefault}
-
-% --- Style -------------------------------------------------------------------
-\renewcommand{\bibfont}{\small \sffamily}
-\renewcommand{\captionfont}{\small\sffamily}
-\renewcommand{\captionlabelfont}{\bfseries}
-
-% --- Section/SubSection/SubSubSection ----------------------------------------
-\titleformat{\section}
-  {\normalfont\sffamily\Large\bfseries}
-  {}{0pt}{}
-\titleformat{\subsection}
-  {\normalfont\sffamily\large\bfseries}
-  {}{0pt}{}
-\titleformat{\subsubsection}
-  {\normalfont\sffamily\bfseries}
-  {}{0pt}{}
-\titleformat*{\paragraph}
-  {\sffamily\normalsize}
-
-
-% --- Header / Footer ---------------------------------------------------------
-\usepackage{fancyhdr}
-\pagestyle{fancy}
-\fancyhf{}
-%\renewcommand{\headrulewidth}{0.50pt}
-\renewcommand{\headrulewidth}{0pt}
-\fancyhead[L]{\hspace{-0.75cm}\includegraphics[width=5.5cm]{logo.png}}
-\fancyhead[C]{}
-\fancyhead[R]{}
-\renewcommand{\footrulewidth}{0.25pt}
-
-\fancyfoot[L]{\footnotesize{\sffamily Pang et al., (2026). HydraR:
-Stateful Agentic Orchestration for Scientific Reproducibility in
-R. \textit{Journal of Open Source
-Software}, 11(121), 7654. \href{https://doi.org/10.21105/joss.07654}{https://doi.org/10.21105/joss.07654}}}
-
-
-\fancyfoot[R]{\sffamily \thepage}
-\makeatletter
-\let\ps@plain\ps@fancy
-\fancyheadoffset[L]{4.5cm}
-\fancyfootoffset[L]{4.5cm}
-
-% --- Macros ---------
-
-\definecolor{linky}{rgb}{0.0, 0.5, 1.0}
-
-\newtcolorbox{repobox}
-   {colback=red, colframe=red!75!black,
-     boxrule=0.5pt, arc=2pt, left=6pt, right=6pt, top=3pt, bottom=3pt}
-
-\newcommand{\ExternalLink}{%
-   \tikz[x=1.2ex, y=1.2ex, baseline=-0.05ex]{%
[…]
