### 2026-04-07 9094ddb
docs: finalize Complete Instruction Manual and synchronize submission release
