+++ vignettes/manual.Rmd
@@ -1,0 +1,168 @@
+---
+title: "HydraR: Complete Instruction Manual"
+author: "Chi Nam Ignatius Pang & Aidan Tay"
+date: "`r Sys.Date()`"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{HydraR: Complete Instruction Manual}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+## 1. Introduction
+
+`HydraR` is a state-managed orchestration framework for building "agentic" workflows in R. Unlike simple chat interfaces, `HydraR` treats AI interactions as components of a larger, verifiable, and reproducible scientific pipeline. 
+
+### Core Philosophy: The Visual Blueprint
+We believe that complex AI logic should be **declarative** and **visual**. By defining your agent's reasoning as a Mermaid diagram, you create a "blueprint" that is:
+- **Shareable**: Domain experts can audit the logic without reading R code.
+- **Auditable**: Every execution path is recorded and can be visualized after the run.
+- **Resilient**: Failures are handled through explicit error edges and model-tiering.
+
+---
+
+## 2. YAML Workflow Anatomy
+
+HydraR workflows are typically defined in `.yml` files. These files provide the high-level orchestration, logic components, and initial state for the Agent Graph.
+
+### Root Keys
+
+| Key | Purpose |
+| :--- | :--- |
+| **`graph`** | A Mermaid string (e.g., `graph TD`) defining the network topology. This is the source of truth for node IDs and edges. |
+| **`conditional_edges`** | Defines branching logic where the next node depends on the output of a previous node (e.g., Success/Failure loops). |
+| **`roles`** | A dictionary of AI personas/system prompts referenced by `role_id`. |
+| **`logic`** | A dictionary of R code blocks (logic functions) or prompt templates. |
+| **`start_node`** | Explicit entry point for the graph (optional, defaults to root nodes). |
+| **`initial_state`** | The seed data injected into the `AgentState` before execution begins. |
+
+---
+
+## 3. Node Technical Gallery
+
+`HydraR` uses specialized R6 classes to handle different types of tasks.
+
+### `type=llm` (AgentLLMNode)
+The primary node for interacting with Large Language Models.
+- **Role**: Combines a system prompt and user input into an LLM call.
+- **Key Parameters**:
+    - `driver`: The execution engine (e.g., `gemini_api`, `ollama`).
+    - `role_id`: Lookup key in the `roles` section.
+
+### `type=logic` (AgentLogicNode)
+Executes deterministic R code.
+- **Role**: Data validation, file I/O, template rendering, and decision gates.
+- **Note**: Must adhere to **APAF Rule G-25** (no `for` loops).
+
+### `type=router` (AgentRouterNode)
+A decision-based node that dynamically selects the next node based on R logic.
+
+### `type=map` (AgentMapNode)
+Iterates over a list in the current state to process multiple items in parallel or sequence.
+
+### `type=observer` (AgentObserverNode)
+Executes logic for side-effects without modifying the main state.
+
+### `type=merge` (MergeHarmonizer)
+Synchronizes parallel execution paths, reconciling changes from isolated git worktrees back into the base state.
+
+---
+
+## 4. Resilience & Execution Control
+
+AI agents are inherently non-deterministic. `HydraR` provides robust tools to manage this uncertainty.
+
+### 4.1 Retries & Timeouts
[…]
