### 2026-04-07 2bca60a
docs: restore hero graphic to man/figures and convert manual to .md
