+++ llms.txt
@@ -4,44 +4,101 @@
 |----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|:---------------------------------------------------------:|
 | [![Lifecycle: stable](https://img.shields.io/badge/lifecycle-stable-green.svg)](https://lifecycle.r-lib.org/articles/stages.html#stable) [![R-CMD-check](https://github.com/apaf-bioinformatics/HydraR/actions/workflows/R-CMD-check.yaml/badge.svg?branch=main)](https://github.com/apaf-bioinformatics/HydraR/actions/workflows/R-CMD-check.yaml) [![Test Coverage](https://github.com/apaf-bioinformatics/HydraR/actions/workflows/test-coverage.yaml/badge.svg)](https://app.codecov.io/gh/apaf-bioinformatics/HydraR) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.19446742.svg)](https://doi.org/10.5281/zenodo.19446742) | ![HydraR Hex Logo](reference/figures/hydrar_hex_logo.png) |
 
-`HydraR` is a lightweight, state-of-the-art orchestrator for building
-general-purpose agentic workflows in R. It prioritizes **CLI-native LLM
-interactions**, **hardened state management**, and **graph-based
-execution** (supporting both Directed Acyclic Graphs and iterative
-loops).
+### *Stateful Agentic Orchestration for Scientific Reproducibility*
 
-## Why HydraR?
+`HydraR` is a lightweight orchestrator for building **autonomous AI
+workflows** natively in R. It allows researchers to design complex,
+multi-step processes—such as data cleaning, analysis, and report
+generation—that are driven by Large Language Models (LLMs) while
+maintaining the rigorous auditability and reproducibility expected in
+scientific research.
+
+![HydraR Visual Blueprint: A sample travel itinerary workflow showing
+iterative validation loops and parallel generation
+paths.](paper/figures/travel_workflow.png)
+
+HydraR Visual Blueprint: A sample travel itinerary workflow showing
+iterative validation loops and parallel generation paths.
+
+## 🎯 What is HydraR? (For Non-Experts)
+
+Think of an **“Agent”** as an AI-powered assistant that can perform
+specific tasks. Instead of just “chatting” with an AI, `HydraR` allows
+you to link multiple assistants together into a **Workflow**.
+
+You create a “blueprint” (a visual map) of your research tasks. `HydraR`
+then manages the execution: it gives each assistant its own isolated
+workspace, remembers every step they take (checkpointing), and allows
+them to “loop back” and fix mistakes if a task fails. This makes AI
+research **reliable, auditable, and automated**.
+
+## ⚖️ Why HydraR? (Statement of Need)
 
 Standard agentic frameworks often rely heavily on brittle API wrappers
-and volatile state. `HydraR` is built for durability and
-reproducibility: - **CLI-First**: Directly drive high-performance CLI
-tools like `gemini-cli`, `claude-code`, or `gh copilot`. - **Hardened
-State**: Implements a robust state machine with persistent checkpointing
-(DuckDB/SQLite). - **Graph-Native**: Design complex logic transitions
-and loops with built-in validation.
[…]
