### 2026-04-07 146fc6e
release: version 0.2.1 - sync case studies with paper and update vignettes
