### 2026-04-07 95da07c
docs: fix stray .Rmd link in README
