### 2026-04-08 115d2ae
chore: fix R CMD check issues (non-portable filenames, build ignores, and vignette layout)
