+++ scripts/prepare_arxiv.sh
@@ -54,7 +54,7 @@ sed -i '' '/\\newcommand{\\rorlogo}/,/}/d' "$CLEAN_TEX"
 # This is specific to the HydraR paper structure
 # We replace the two figure environments with subfigures
 # NOTE: This assumes the JOSS output format remains consistent with version 1.5.x
-perl -0777 -i -pe 's/\\begin\{figure\}.*?fig:travel_workflow_a\}.*?\\end\{figure\}\s*\\begin\{figure\}.*?fig:travel_workflow_b\}.*?\\end\{figure\}/\\begin{figure}\n\\centering\n\\begin{subfigure}[b]{0.48\\textwidth}\n\\centering\n\\pandocbounded{\\includegraphics[width=\\textwidth]{figures\/Itinerary Page 1.png}}\n\\caption{Travel pamphlet visualization.}\n\\label{fig:travel_workflow_a}\n\\end{subfigure}\n\\hfill\n\\begin{subfigure}[b]{0.48\\textwidth}\n\\centering\n\\pandocbounded{\\includegraphics[width=\\textwidth]{figures\/Itinerary Page 2.png}}\n\\caption{Final CSS-styled itinerary.}\n\\label{fig:travel_workflow_b}\n\\end{subfigure}\n\\caption{Multi-modal travel itinerary planner results: (a) generated visual content; (b) final formatted itinerary.}\n\\label{fig:travel_workflow}\n\\end{figure}/sg' "$CLEAN_TEX"
+perl -0777 -i -pe 's/\\begin\{figure\}.*?fig:travel_workflow_a\}.*?\\end\{figure\}\s*\\begin\{figure\}.*?fig:travel_workflow_b\}.*?\\end\{figure\}/\\begin{figure}\n\\centering\n\\begin{subfigure}[b]{0.48\\textwidth}\n\\centering\n\\pandocbounded{\\includegraphics[width=\\textwidth]{figures\/itinerary_page_1.png}}\n\\caption{Travel pamphlet visualization.}\n\\label{fig:travel_workflow_a}\n\\end{subfigure}\n\\hfill\n\\begin{subfigure}[b]{0.48\\textwidth}\n\\centering\n\\pandocbounded{\\includegraphics[width=\\textwidth]{figures\/itinerary_page_2.png}}\n\\caption{Final CSS-styled itinerary.}\n\\label{fig:travel_workflow_b}\n\\end{subfigure}\n\\caption{Multi-modal travel itinerary planner results: (a) generated visual content; (b) final formatted itinerary.}\n\\label{fig:travel_workflow}\n\\end{figure}/sg' "$CLEAN_TEX"
 
 # Copy figures to arxiv folder
 cp -r "$MANUSCRIPT_DIR/figures/"* "$ARXIV_DIR/figures/"

[…]
