+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   sorting_benchmark: sorting_benchmark.html
   state_persistence: state_persistence.html
   targets_integration: targets_integration.html
-last_built: 2026-04-07T17:36Z
+last_built: 2026-04-08T03:06Z
 urls:
   reference: https://github.com/APAF-bioinformatics/HydraR/reference
   article: https://github.com/APAF-bioinformatics/HydraR/articles
