### 2026-04-08 795b8cc
Finalize v0.2.2: Add Acknowledgements, NATA Accreditation, and Resolve CLI test failures
