### 2026-04-08 d193ace
style: automated code styling with styler
