+++ references.md
@@ -14,6 +14,8 @@ standards.
 ## Project Resources
 
 - HydraR Documentation: <https://apaf-bioinformatics.github.io/HydraR/>
+- HydraR v0.2.2 Release:
+  <https://github.com/APAF-bioinformatics/HydraR/releases/tag/v0.2.2>
 
 ## Package Dependencies
 
