### 2026-04-08 076b3d4
Finalize HydraR v0.2.1: Add formal Acknowledgements, NATA accreditation, and institutional branding
