### 2026-04-09 #766 @IgnatiusPang
Thank you editor for the checks. I will work on the list of errors e.g. the list of functions without documented return values and the list of functions without examples.

I will then progress to v0.2.2 and post an update here.

Would you please let me know what I have to do once I have fixed these problems and updated to v0.2.2 to continue the peer review?
