### 2026-04-09 8f6d1ac
style: address goodpractice warnings (cyclocomplexity, line lengths, sapply)
