+++ scripts/inject_roxygen.py
@@ -1,0 +1,329 @@
+import os
+import re
+
+updates = {
+    "AgentBashNode": {
+        "return": "#' @return An `AgentBashNode` object.\n",
+        "example": "#' @examples\n#' \\dontrun{\n#' node <- AgentBashNode$new(id = \"bash_1\")\n#' node$call(\"echo hello\")\n#' }\n"
+    },
+    "AgentPythonNode": {
+        "return": "#' @return An `AgentPythonNode` object.\n",
+        "example": "#' @examples\n#' \\dontrun{\n#' node <- AgentPythonNode$new(id = \"py_1\")\n#' node$call(\"print('hello')\")\n#' }\n"
+    },
+    "AgentTool": {
+        "return": "#' @return An `AgentTool` object.\n",
+        "example": "#' @examples\n#' \\dontrun{\n#' tool <- AgentTool$new(name = \"my_tool\", description = \"A tool\")\n#' }\n"
+    },
+    "AnthropicAPIDriver": {
+        "return": "#' @return An `AnthropicAPIDriver` object.\n",
+        "example": "#' @examples\n#' \\dontrun{\n#' driver <- AnthropicAPIDriver$new()\n#' driver$call(\"Hello, Anthropic\")\n#' }\n"
+    },
+    "AnthropicCLIDriver": {
+        "return": "#' @return An `AnthropicCLIDriver` object.\n",
+        "example": "#' @examples\n#' \\dontrun{\n#' driver <- AnthropicCLIDriver$new()\n#' driver$call(\"Hello, Claude\")\n#' }\n"
+    },
+    "Checkpointer": {
+        "return": "#' @return A `Checkpointer` object.\n",
+        "example": "#' @examples\n#' \\dontrun{\n#' cp <- Checkpointer$new(saver = MemorySaver$new())\n#' }\n"
+    },
+    "CopilotCLIDriver": {
+        "return": "#' @return A `CopilotCLIDriver` object.\n",
+        "example": "#' @examples\n#' \\dontrun{\n#' driver <- CopilotCLIDriver$new()\n#' driver$call(\"Write a simple function\")\n#' }\n"
+    },
+    "DuckDBMessageLog": {
+        "return": "#' @return A `DuckDBMessageLog` object.\n",
+        "example": "#' @examples\n#' \\dontrun{\n#' log <- DuckDBMessageLog$new(session_id = \"test\")\n#' }\n"
+    },
+    "JSONLMessageLog": {
+        "return": "#' @return A `JSONLMessageLog` object.\n",
+        "example": "#' @examples\n#' \\dontrun{\n#' log <- JSONLMessageLog$new(session_id = \"test\")\n#' }\n"
+    },
+    "GeminiCLIDriver": {
+        "return": "#' @return A `GeminiCLIDriver` object.\n",
+        "example": "#' @examples\n#' \\dontrun{\n#' driver <- GeminiCLIDriver$new()\n#' driver$call(\"Hello, Gemini\")\n#' }\n"
+    },
+    "MemoryMessageLog": {
+        "return": "#' @return A `MemoryMessageLog` object.\n",
+        "example": "#' @examples\n#' \\dontrun{\n#' log <- MemoryMessageLog$new()\n#' }\n"
+    },
+    "MessageLog": {
+        "return": "#' @return A `MessageLog` base object.\n",
+        "example": "#' @examples\n#' \\dontrun{\n#' # This is an abstract base class.\n#' # Instantiate a subclass like MemoryMessageLog.\n#' }\n"
+    },
+    "MemorySaver": {
+        "return": "#' @return A `MemorySaver` object.\n",
[…]
