+++ reference/MemorySaver.md
@@ -3,6 +3,10 @@
 In-memory implementation of the Checkpointer interface. Stores
 checkpoints in an R environment.
 
+## Value
+
+A \`MemorySaver\` object.
+
 ## Super class
 
 [`HydraR::Checkpointer`](https://github.com/APAF-bioinformatics/HydraR/reference/Checkpointer.md)
@@ -93,3 +97,11 @@ The objects of this class are cloneable with this method.
 - `deep`:
 
   Whether to make a deep clone.
+
+## Examples
+
+``` r
+if (FALSE) { # \dontrun{
+saver <- MemorySaver$new()
+} # }
+```
