+++ reference/ConflictResolver.md
@@ -96,3 +96,11 @@ The objects of this class are cloneable with this method.
 - `deep`:
 
   Whether to make a deep clone.
+
+## Examples
+
+``` r
+if (FALSE) { # \dontrun{
+resolver <- ConflictResolver$new(worktree_dir = ".")
+} # }
+```
