+++ reference/AgentObserverNode.md
@@ -93,3 +93,11 @@ The objects of this class are cloneable with this method.
 - `deep`:
 
   Whether to make a deep clone.
+
+## Examples
+
+``` r
+if (FALSE) { # \dontrun{
+node <- AgentObserverNode$new(id = "obs1", observer_fn = function(s) print(s))
+} # }
+```
