### 2026-04-09 8cd57be
docs: complete comprehensive documentation overhaul of core components
