+++ R/checkpointer.R
@@ -9,28 +9,35 @@
 #' Checkpointer Interface
 #'
 #' @description
-#' Abstract base class for AgentDAG checkpointers.
+#' An abstract base class defining the contract for state persistence and
+#' recovery in HydraR. Checkpointers allow a DAG to be paused and resumed
+#' across sessions by saving the \code{AgentState} after each node execution.
 #'
 #' @importFrom R6 R6Class
-#' @return A `Checkpointer` object.
+#' @return A \code{Checkpointer} object.
+#'
 #' @examples
 #' \dontrun{
-#' cp <- Checkpointer$new(saver = MemorySaver$new())
+#' # Checkpointers are used within AgentDAG$run()
+#' # Use a concrete implementation like RDSSaver or DuckDBSaver.
 #' }
 #' @export
 Checkpointer <- R6::R6Class("Checkpointer",
   public = list(
     #' @description Persist state to the checkpointer.
-    #' @param thread_id String. Identifier for the execution thread.
-    #' @param state AgentState object. The state to save.
+    #' @param thread_id String.
+    #' A unique identifier for the execution thread or session.
+    #' @param state AgentState.
+    #' The \code{AgentState} object to be persisted.
     #' @return NULL (called for side effect).
     put = function(thread_id, state) {
       stop("Method 'put' must be implemented by subclass.")
     },
 
-    #' Load state
-    #' @param thread_id String. Identifier for the execution thread.
-    #' @return AgentState object or NULL if not found.
+    #' Load state from the checkpointer
+    #' @param thread_id String.
+    #' The unique identifier associated with the saved state.
+    #' @return AgentState object or NULL if no checkpoint is found for the thread.
     get = function(thread_id) {
       stop("Method 'get' must be implemented by subclass.")
     }
@@ -40,14 +47,19 @@ Checkpointer <- R6::R6Class("Checkpointer",
 #' MemorySaver Checkpointer
 #'
 #' @description
-#' In-memory implementation of the Checkpointer interface.
-#' Stores checkpoints in an R environment.
+#' An in-memory implementation of the \code{Checkpointer} interface. Stores
+#' checkpoints in a dedicated R environment. This is useful for testing or
+#' short-lived sessions where persistence to disk is not required.
 #'
 #' @importFrom R6 R6Class
-#' @return A `MemorySaver` object.
+#' @return A \code{MemorySaver} object.
+#'
 #' @examples
 #' \dontrun{
 #' saver <- MemorySaver$new()
+#' dag <- dag_create(checkpointer = saver)
+#'
+#' # State is saved in saver$storage environment
 #' }
 #' @export
 MemorySaver <- R6::R6Class("MemorySaver",
@@ -106,15 +118,21 @@ MemorySaver <- R6::R6Class("MemorySaver",
 #' RDS File Checkpointer
 #'
 #' @description
-#' Lightweight file-based checkpointer using base R `saveRDS`/`readRDS`.
-#' Each thread is persisted as a separate `.rds` file in the specified directory.
-#' No external dependencies required.
+#' A lightweight, file-based checkpointer that uses R's native \code{saveRDS}
+#' and \code{readRDS} functions. Each thread is saved as an individual \code{.rds}
+#' file in a specified directory.
 #'
 #' @importFrom R6 R6Class
-#' @return An `RDSSaver` object.
+#' @return An \code{RDSSaver} object.
+#'
 #' @examples
 #' \dontrun{
-#' saver <- RDSSaver$new()
+#' # Save checkpoints to a local directory
+#' saver <- RDSSaver$new(dir = "my_checkpoints")
+#'
[…]
