+++ R/dag.R
@@ -10,29 +10,51 @@
 #' Agent Graph R6 Class
 #'
 #' @description
-#' Defines and executes a Directed Graph of AgentNodes.
-#' Supports both pure DAG execution (parallel) and iterative loops via conditional edges.
+#' The core orchestrator in HydraR, \code{AgentDAG} defines and executes a
+#' Directed Graph of \code{AgentNode} objects. It supports:
+#' \itemize{
+#'   \item \strong{Pure DAGs}: Parallel execution using \code{furrr}.
+#'   \item \strong{Cycles & Loops}: Iterative execution via conditional edges.
+#'   \item \strong{State Isolation}: Parallel branch execution in isolated git worktrees.
+#'   \item \strong{Persistence}: Automatic state checkpointing and restoration.
+#' }
 #'
-#' @return An `AgentDAG` R6 object.
+#' @return An \code{AgentDAG} R6 object.
+#'
 #' @examples
+#' \dontrun{
+#' # Define a simple linear workflow
 #' dag <- AgentDAG$new()
-#' node <- AgentLogicNode$new("start", function(state) list(status = "success"))
-#' dag$add_node(node)
+#'
+#' # Define a logic function
+#' fetch_data <- function(state) {
+#'   list(status = "success", output = list(raw = "Some data"))
+#' }
+#'
+#' # Add nodes
+#' dag$add_node(AgentLogicNode$new("fetcher", logic_fn = fetch_data))
+#' dag_add_llm_node(
+#'   dag,
+#'   id = "analyzer",
+#'   role = "Analyze the data in the state.",
+#'   driver = GeminiCLIDriver$new()
+#' )
+#'
+#' # Connect nodes
+#' dag$add_edge("fetcher", "analyzer")
+#'
+#' # Compile and verify the graph
+#' dag$compile()
+#'
+#' # Execution requires an initial state
+#' results <- dag$run(initial_state = list(input_file = "raw.txt"))
+#' }
 #' @importFrom R6 R6Class
 #' @importFrom digest digest
 #' @importFrom jsonlite toJSON fromJSON write_json
 #' @importFrom igraph graph_from_data_frame is_dag topo_sort edge V vcount make_empty_graph degree is_connected components bfs
 #' @importFrom furrr future_map
 #' @importFrom purrr map set_names iwalk walk
-#' @examples
-#' \dontrun{
-#' dag <- AgentDAG$new()
-#' dag$add_node(AgentNode$new("A"))
-#' dag$add_node(AgentNode$new("B"))
-#' dag$add_edge("A", "B")
-#' dag$set_start_node("A")
-#' dag$compile()
-#' }
 #' @export
 AgentDAG <- R6::R6Class("AgentDAG",
   public = list(
@@ -60,6 +82,7 @@ AgentDAG <- R6::R6Class("AgentDAG",
     worktree_manager = NULL,
 
     #' @description Initialize AgentDAG
+    #' @return A new \code{AgentDAG} instance with empty node and edge lists.
     initialize = function() {
       self$nodes <- list()
       self$edges <- list()
@@ -74,9 +97,16 @@ AgentDAG <- R6::R6Class("AgentDAG",
       self$worktree_manager <- NULL
     },
 
-    #' Set Start Node(s)
-    #' @param node_ids Character vector of node IDs.
-    #' @return The AgentDAG object (invisibly).
+    #' @description Set Start Node(s)
+    #' Explicitly defines the entry point(s) of the graph. If not called, the
+    #' engine defaults to nodes with an in-degree of 0. Required for graphs
+    #' with cycles or where execution must start at a specific node.
+    #'
+    #' @param node_ids Character vector.
+    #' A vector of one or more node IDs. These nodes must already exist in the
[…]
