+++ reference/dag_add_llm_node.md
@@ -1,6 +1,7 @@
 # Add an LLM Agent Node directly to a DAG
 
-Add an LLM Agent Node directly to a DAG
+Instantiates an `AgentLLMNode` and appends it to the provided `AgentDAG`
+in one step.
 
 ## Usage
 
@@ -12,7 +13,7 @@ dag_add_llm_node(dag, id, role, driver, model = NULL, cli_opts = list(), ...)
 
 - dag:
 
-  AgentDAG object.
+  AgentDAG. The graph object to which the node will be added.
 
 - id:
 
@@ -24,29 +25,36 @@ dag_add_llm_node(dag, id, role, driver, model = NULL, cli_opts = list(), ...)
 
 - driver:
 
-  AgentDriver object.
+  AgentDriver. The LLM driver instance.
 
 - model:
 
-  String. Optional model override.
+  String. Optional model name override.
 
 - cli_opts:
 
-  List. Optional CLI options.
+  List. Optional CLI/API parameters.
 
 - ...:
 
-  Additional arguments passed to AgentLLMNode\$new()
+  Additional arguments. Passed to `AgentLLMNode$new()`.
 
 ## Value
 
-The modified AgentDAG object (invisibly).
+The modified `AgentDAG` object (invisibly).
 
 ## Examples
 
 ``` r
 if (FALSE) { # \dontrun{
+# NOTE: Set GOOGLE_API_KEY in your .Renviron for Gemini drivers.
+
 dag <- dag_create()
-dag <- dag_add_llm_node(dag, "node1", "Assistant", AnthropicAPIDriver$new())
+dag_add_llm_node(
+  dag,
+  id = "summary_node",
+  role = "Summarise the following text.",
+  driver = GeminiAPIDriver$new()
+)
 } # }
 ```
