+++ reference/OpenAICodexCLIDriver.md
@@ -1,10 +1,12 @@
 # OpenAI Codex CLI Driver R6 Class
 
-Driver for the official \`codex\` CLI tool (v0.118+).
+A specialized `AgentDriver` for the official `codex` CLI tool (v0.118+).
+This model is legacy but still supported for specific code-generation
+tasks.
 
 ## Value
 
-An \`OpenAICodexCLIDriver\` object.
+An `OpenAICodexCLIDriver` object.
 
 ## Super class
 
@@ -110,7 +112,9 @@ The objects of this class are cloneable with this method.
 
 ``` r
 if (FALSE) { # \dontrun{
+# Ensure Codex CLI is installed and authenticated
 driver <- OpenAICodexCLIDriver$new()
-driver$call("Hello, Codex")
+
+response <- driver$call("Create a data.frame with 5 rows and 2 columns.")
 } # }
 ```
