### 2026-04-09 a9eb321
docs: finalize v0.2.2 release with NEWS.md, test hygiene, and documentation polish
