+++ tests/testthat/test-story_teller.R
@@ -69,13 +69,13 @@ test_that("Story Teller collaboration works", {
     if_false = "Writer" # Loop back
   )
 
-  capture_warnings(dag$compile())
+  suppressWarnings(dag$compile())
   compiled_dag <- dag
   # Run the DAG
-  result <- compiled_dag$run(
+  result <- suppressWarnings(compiled_dag$run(
     initial_state = list(story_prompt = "A robot learning to cook."),
     max_steps = 10
-  )
+  ))
   # Assertions
   expect_equal(loop_count, 2)
   expect_equal(result$state$get("Reviewer"), "Approved")
