### 2026-04-09 0dff1ae
docs: rewrite manual.md as comprehensive narrative-style beginner guide
