+++ vignettes/manual.md
@@ -1,146 +1,827 @@
-# HydraR: Complete Instruction Manual
+# HydraR: The Complete Guide
 
-## 1. Introduction
+## Welcome
 
-`HydraR` is a state-managed orchestration framework for building "agentic" workflows in R. Unlike simple chat interfaces, `HydraR` treats AI interactions as components of a larger, verifiable, and reproducible scientific pipeline. 
+If you've ever wished you could hand off a tedious, multi-step research task to an AI assistant—but needed the results to be *reproducible*, *auditable*, and *safe*—then HydraR is the tool you've been looking for.
 
-### Core Philosophy: The Visual Blueprint
-We believe that complex AI logic should be **declarative** and **visual**. By defining your agent's reasoning as a Mermaid diagram, you create a "blueprint" that is:
-- **Shareable**: Domain experts can audit the logic without reading R code.
-- **Auditable**: Every execution path is recorded and can be visualized after the run.
-- **Resilient**: Failures are handled through explicit error edges and model-tiering.
+This guide will walk you through HydraR from the ground up. We'll start with the simplest possible example—a single node that says "hello"—and build our way up to workflows that loop, branch, checkpoint, and even run agents in isolated sandboxes. By the end, you'll be comfortable building your own agentic pipelines in R.
+
+No prior experience with AI orchestration is needed. If you can write an R function and read a flowchart, you're ready.
 
 ---
 
-## 2. YAML Workflow Anatomy
+## Part 1: What Is HydraR?
 
-HydraR workflows are typically defined in `.yml` files. These files provide the high-level orchestration, logic components, and initial state for the Agent Graph.
+### The Problem
 
-### Root Keys
+Imagine you're a researcher who needs to:
 
-| Key | Purpose |
-| :--- | :--- |
-| **`graph`** | A Mermaid string (e.g., `graph TD`) defining the network topology. This is the source of truth for node IDs and edges. |
-| **`conditional_edges`** | Defines branching logic where the next node depends on the output of a previous node (e.g., Success/Failure loops). |
-| **`roles`** | A dictionary of AI personas/system prompts referenced by `role_id`. |
-| **`logic`** | A dictionary of R code blocks (logic functions) or prompt templates. |
-| **`start_node`** | Explicit entry point for the graph (optional, defaults to root nodes). |
-| **`initial_state`** | The seed data injected into the `AgentState` before execution begins. |
+1. Pull data from a database.
+2. Ask an LLM to summarise the key findings.
+3. Have a second LLM critically review that summary.
+4. If the review fails, loop back and try again.
+5. Save a checkpoint after every step, so a network failure at step 3 doesn't waste the work from steps 1 and 2.
 
----
+You could write a long R script with nested `tryCatch` blocks and manual state tracking, but that approach quickly becomes unreadable, fragile, and impossible to share with collaborators who aren't programmers. HydraR solves this by letting you describe your workflow as a **visual graph**—a flowchart—and then executing it with full state management, persistence, and error recovery built in.
 
-## 3. Node Technical Gallery
+### The Core Idea
 
-`HydraR` uses specialized R6 classes to handle different types of tasks.
+HydraR is built around three concepts:
 
-### `type=llm` (AgentLLMNode)
-The primary node for interacting with Large Language Models.
-- **Role**: Combines a system prompt and user input into an LLM call.
[…]
