### 2026-04-09 223681a
docs: rewrite Part 7 with direct side-by-side comparison showing YAML advantage
