+++ vignettes/manual.md
@@ -294,73 +294,109 @@ cat("\nLLM was called", driver$call_count, "times")
 
 ## Part 7: The YAML-First Approach (Declarative Workflows)
 
-Writing R code to build a DAG is perfectly fine for development, but for sharing, auditing, and version-controlling your workflows, HydraR offers a **declarative** alternative: YAML files with embedded Mermaid diagrams.
+Look back at Part 6. Building that Generate→Validate→Loop workflow took about 40 lines of R code, and that was with a mock driver. With a real driver and more realistic logic, it could easily be 60–80 lines. Now imagine handing that script to a collaborator who isn't an R programmer and asking them to audit the logic. It would be a tough conversation.
 
-### Why YAML?
+HydraR offers a radically simpler alternative: describe the **same** workflow in a YAML file, and let `spawn_dag()` build it for you.
 
-- **Readable by non-programmers**: A domain expert can audit the workflow structure without knowing R.
-- **Version-controllable**: YAML files produce clean, meaningful diffs in Git.
-- **Portable**: The same YAML file can be loaded in any R environment with HydraR installed.
+### The Same Workflow, In YAML
 
-### Anatomy of a Workflow File
+Here is the exact same Generate→Validate→Loop pattern from Part 6, written declaratively:
 
 ```yaml
-# my_workflow.yml
+# question_answerer.yml
 
-# The graph defines the topology using Mermaid syntax.
-# Node labels encode the type and parameters using pipe-separated key=value pairs.
 graph: |
   graph TD
-    Analyst["Research Analyst | type=llm | role_id=analyst_role"]
-    Reviewer["Quality Check | type=logic | logic_id=review_fn"]
-    Analyst --> Reviewer
-    Reviewer -- "fail" --> Analyst
+    Generator["Answer Generator | type=llm | role_id=answerer"]
+    Validator["Quality Gate | type=logic | logic_id=check_answer"]
+    Generator --> Validator
+    Validator -- "retry" --> Generator
 
-# Roles define the system prompts for LLM nodes.
 roles:
-  analyst_role: >
-    You are a senior research analyst. Provide concise, evidence-based
-    summaries of the data presented to you.
+  answerer: >
+    Answer the user's question concisely. Always include
+    the number 42 in your final answer.
 
-# Logic blocks define the R functions used by logic nodes.
-# These can be inline R code or paths to .R files.
 logic:
-  review_fn: |
+  check_answer: |
     function(state) {
-      output <- state$get("Analyst")
-      has_evidence <- grepl("evidence|data|study", output, ignore.case = TRUE)
-      list(status = "success", output = list(approved = has_evidence))
+      answer <- state$get("Generator")
+      is_good <- grepl("42", answer)
+      list(status = "success", output = list(passed = is_good))
     }
 
-# Conditional edges define the branching logic.
 conditional_edges:
-  Reviewer:
+  Validator:
     test: |
-      function(result) isTRUE(result$approved)
-    if_true: ~     # ~ means NULL (stop)
-    if_false: Analyst
+      function(result) isTRUE(result$passed)
+    if_true: ~          # ~ means NULL (stop — we're done)
+    if_false: Generator  # loop back and try again
 
-# Initial state seeds the workflow with starting data.
 initial_state:
-  research_question: "What are the effects of sleep deprivation on cognition?"
+  question: "What is the meaning of life?"
 ```
 
-### Loading and Running
+And the R code to run it:
 
 ```r
 library(HydraR)
 
-# Load the YAML file — this registers all roles and logic automatically
-wf <- load_workflow("my_workflow.yml")
+wf      <- load_workflow("question_answerer.yml")
+dag     <- spawn_dag(wf)
[…]
