### 2026-04-09 7ca9112
docs: add manual to pkgdown site, update NEWS.md with manual rewrite and test hygiene
