+++ NEWS.md
@@ -13,6 +13,8 @@
     - Fixed Roxygen parsing errors where method titles merged into descriptions in `man/AgentDAG.Rd`.
     - Restored flexible function lookup in `auto_node_factory` to maintain backward compatibility with global functions in tests.
 - **Clean Registry**: Package now achieves a perfectly clean `R CMD check` (0 Errors, 0 Warnings, 0 Notes) on core builds.
+- **Complete Instruction Manual**: Rewrote `vignettes/manual.md` as a 16-part narrative-style beginner guide, progressing from a single "hello world" node through loops, YAML workflows, checkpointing, and git worktree isolation. Includes a direct side-by-side comparison showing the YAML-first approach reduces a 40-line R script to 4 lines.
+- **Test Hygiene**: Silenced expected diagnostic warnings across the test suite for cyclic and multi-root graph patterns. Removed deprecated `testthat::context()` calls.
 
 # HydraR 0.2.1 (Final Polish & Acknowledgements)
 
