### 2026-04-09 3143849
Final documentation overhaul and pkgdown site build for manual v0.2.2
