+++ R/tools.R
@@ -16,7 +16,19 @@
 #' @return An `AgentTool` object.
 #' @examples
 #' \dontrun{
-#' tool <- AgentTool$new(name = "my_tool", description = "A tool")
+#' # Define a tool for searching genomic databases
+#' tool <- AgentTool$new(
+#'   name = "blast_search",
+#'   description = "Perform a BLAST search against the NCBI non-redundant database.",
+#'   parameters = list(
+#'     query = "The DNA sequence string",
+#'     evalue = "The e-value threshold (default 1e-5)"
+#'   ),
+#'   example = "blast_search(query='ATGC...', evalue=0.001)"
+#' )
+#'
+#' # Format for injection into a system prompt
+#' message(tool$format())
 #' }
 #' @export
 AgentTool <- R6::R6Class("AgentTool",
@@ -70,7 +82,14 @@ AgentTool <- R6::R6Class("AgentTool",
 #' @return A formatted string containing all tool descriptions.
 #' @examples
 #' \dontrun{
-#' format_toolset(list(my_tool = AgentTool$new(...)))
+#' # Format a collection of tools for an agent
+#' tools <- list(
+#'   search = AgentTool$new("google_search", "Search the web"),
+#'   run_r = AgentTool$new("r_exec", "Execute R code locally")
+#' )
+#'
+#' prompt_appendix <- format_toolset(tools)
+#' cat(prompt_appendix)
 #' }
 #' @export
 format_toolset <- function(tools) {

[…]
