+++ reference/standard_node_factory.md
@@ -31,6 +31,11 @@ AgentNode object.
 
 ``` r
 if (FALSE) { # \dontrun{
-node <- standard_node_factory("id", "label")
+# 1. Fast resolve from a Mermaid string with defaults
+drv <- resolve_default_driver("gemini")
+node <- standard_node_factory("n1", "logic:clean_data")
+
+# 2. Using the LLM type with a provided driver
+node_llm <- standard_node_factory("n2", "llm:Technical Writer", driver = drv)
 } # }
 ```
