+++ reference/ConflictResolver.md
@@ -101,6 +101,18 @@ The objects of this class are cloneable with this method.
 
 ``` r
 if (FALSE) { # \dontrun{
-resolver <- ConflictResolver$new(worktree_dir = ".")
+# 1. Initialize a resolver with an LLM strategy
+resolver <- ConflictResolver$new(
+  strategy = "llm",
+  driver = AnthropicAPIDriver$new(model = "claude-3-sonnet")
+)
+
+# 2. Resolve a detected conflict during a merge
+res <- resolver$resolve(
+  repo_root = ".",
+  branch_a = "feature-updates",
+  branch_b = "main",
+  files = c("R/core.R")
+)
 } # }
 ```
