+++ reference/RDSSaver.md
@@ -33,7 +33,7 @@ An `RDSSaver` object.
 
 ------------------------------------------------------------------------
 
-### Method `new()`
+### Method [`new()`](https://rdrr.io/r/methods/new.html)
 
 Initialize RDSSaver
 
@@ -107,11 +107,14 @@ The objects of this class are cloneable with this method.
 
 ``` r
 if (FALSE) { # \dontrun{
-# Save checkpoints to a local directory
-saver <- RDSSaver$new(dir = "my_checkpoints")
+# 1. Persistent checkpointing to a local directory
+saver <- RDSSaver$new(dir = "vault/checkpoints")
 
-# Later, resume execution using the same thread_id
+# 2. Create a DAG and run it with a specific thread ID
 dag <- dag_create(checkpointer = saver)
-dag$run(thread_id = "session_001")
+dag$run(thread_id = "agent_session_alpha")
+
+# 3. Later, resume the same session - HydraR will load the RDS file
+dag$run(thread_id = "agent_session_alpha")
 } # }
 ```
