+++ reference/reducer_append.md
@@ -28,10 +28,15 @@ A combined vector or list containing both `current` and `new`.
 
 ``` r
 if (FALSE) { # \dontrun{
-# Used in AgentState initialization
+# Accumulate a trace of agent IDs
 state <- AgentState$new(
-  initial_data = list(logs = list()),
-  reducers = list(logs = reducer_append)
+  initial_data = list(visited = character()),
+  reducers = list(visited = reducer_append)
 )
+
+state$update(list(visited = "agent_a"))
+state$update(list(visited = "agent_b"))
+
+print(state$get("visited")) # [1] "agent_a" "agent_b"
 } # }
 ```
