+++ R/dag.R
@@ -28,9 +28,9 @@
 #' dag <- AgentDAG$new()
 #'
 #' # Define logic for data retrieval and quality control
-#' register_logic("fetcher", function(state) list(status="success", output="raw data"))
+#' register_logic("fetcher", function(state) list(status = "success", output = "raw data"))
 #' register_logic("gate", function(state) {
-#'   if (nchar(state$get("fetcher")) > 5) list(status="true") else list(status="false")
+#'   if (nchar(state$get("fetcher")) > 5) list(status = "true") else list(status = "false")
 #' })
 #'
 #' # 2. Programmatic Node Construction
@@ -40,8 +40,10 @@
 #' # 3. Connecting nodes with status-based edges
 #' # Errors in 'start' go to a terminal cleanup node
 #' dag$add_edge("start", "check")
-#' dag$add_conditional_edge("check", test = function(res) res$status == "true",
-#'                          if_true = "process", if_false = "retry")
+#' dag$add_conditional_edge("check",
+#'   test = function(res) res$status == "true",
+#'   if_true = "process", if_false = "retry"
+#' )
 #'
 #' # 4. Multi-agent execution context
 #' # Compiling and running with initial state
