### 2026-04-09 #766 @IgnatiusPang
Hi @emilyriederer,

I have implemented the fixes from the Editor Check phase and updated the package to v0.2.2.

Summary of Updates:

* Documentation: Added missing @return values and descriptive @examples for all core R6 methods and functions.
* NEWS: Detailed changes are available in the [v0.2.2 release notes](https://github.com/APAF-bioinformatics/HydraR/blob/main/NEWS.md).

The package is now ready for a re-audit and the next stage of review.
