### 2026-04-09 #766 @mpadge
Thanks @IgnatiusPang. It might help you to install our [`pkgcheck-action`](https://github.com/ropensci-review-tools/pkgcheck-action) in your own repo to confirm, as recommended in [our _Dev Guide_](https://devguide.ropensci.org/softwarereview_author.html#automatic-checks). When you're ready, you can trigger the checks yourself exactly as [@emilyriederer did above](https://github.com/ropensci/software-review/issues/766#issuecomment-4210997479). Thanks 😄
