### 2026-04-09 #766 @IgnatiusPang
Since there was something wrong with the Package Dependencies report and Network Visualisation report, I tried to fix whatever formatting error that could be causing that problem and move to new version v0.2.3. Resubmitting this version and triggering the package check:
