### 2026-04-09 48f4e8b
release: v0.2.3 - Pkgstats Network Recovery
