+++ .github/workflows/pkgdown.yaml
@@ -33,6 +33,12 @@ jobs:
           extra-packages: any::pkgdown, local::.
           needs: website
 
+      - name: Run Tests 🧪
+        run: |
+          install.packages("testthat")
+          testthat::test_local()
+        shell: Rscript {0}
+
       - name: Build site
         run: pkgdown::build_site_github_pages(new_process = FALSE, install = FALSE)
         shell: Rscript {0}
