+++ DESCRIPTION
@@ -1,95 +1,36 @@
 Package: HydraR
 Type: Package
 Title: Stateful Agentic Orchestration for Scientific Reproducibility
-Version: 0.2.2
-Authors@R: 
-    c(person(given = c("Chi", "Nam", "Ignatius"),
-             family = "Pang",
-             role = c("aut", "cre"),
-             email = "ignatius.pang@mq.edu.au",
-             comment = c(ORCID = "0000-0001-9703-5741")),
-      person(given = "Aidan",
-             family = "Tay",
-             role = "aut",
-             comment = c(ORCID = "0000-0003-1315-4896")),
-      person(given = c("APAF", "Agentic"),
-             family = "Workflow",
-             role = "ctb"))
-Description: A high-performance framework for orchestrating complex 
-    agentic workflows in R, specifically designed for scientific 
-    reproducibility and auditability. HydraR provides a robust, 
-    state-managed engine for Directed Acyclic Graphs (DAGs) and 
-    iterative state machines, prioritizing CLI-native LLM 
-    interactions (e.g., Gemini, Claude, Copilot) and hardened 
-    state persistence (DuckDB/SQLite). It features isolated 
-    execution via Git Worktrees for parallel file-modifying 
-    tasks, autonomous quality control through integrated 
-    auditors, and human-readable visualization using Mermaid.js. 
-    Ideal for creating auditable, resumable research assistants 
-    and complex bioinformatics pipelines.
+Version: 0.2.3
+Authors@R: c(person(given = c("Chi", "Nam", "Ignatius"), family =
+        "Pang", role = c("aut", "cre"), email =
+        "ignatius.pang@mq.edu.au", comment = c(ORCID =
+        "0000-0001-9703-5741")), person(given = "Aidan", family =
+        "Tay", role = "aut", comment = c(ORCID =
+        "0000-0003-1315-4896")), person(given = c("APAF", "Agentic"),
+        family = "Workflow", role = "ctb"))
+Description: A high-performance framework for orchestrating complex
+        agentic workflows in R, specifically designed for scientific
+        reproducibility and auditability. HydraR provides a robust,
+        state-managed engine for Directed Acyclic Graphs (DAGs) and
+        iterative state machines, prioritizing CLI-native LLM
+        interactions (e.g., Gemini, Claude, Copilot) and hardened state
+        persistence (DuckDB/SQLite). It features isolated execution via
+        Git Worktrees for parallel file-modifying tasks, autonomous
+        quality control through integrated auditors, and human-readable
+        visualization using Mermaid.js. Ideal for creating auditable,
+        resumable research assistants and complex bioinformatics
+        pipelines.
 URL: https://github.com/APAF-bioinformatics/HydraR
 BugReports: https://github.com/APAF-bioinformatics/HydraR/issues
 License: LGPL (>= 3)
 Encoding: UTF-8
 RoxygenNote: 7.3.3
-Depends: 
-    R (>= 4.1.0)
-Imports: 
-    uuid,
-    DBI,
-    digest,
-    furrr,
-    igraph,
-    jsonlite,
-    purrr,
-    R6,
-    yaml,
-    httr2,
-    base64enc
-Suggests: 
-    testthat (>= 3.0.0),
-    duckdb,
-    knitr,
-    rmarkdown,
-    future,
-    devtools,
-    withr,
-    reticulate,
-    DiagrammeR,
-    ggplot2,
-    DiagrammeRsvg,
-    lintr,
[…]
