### 2026-04-09 146ea9f
fix: mermaid parser empty label logic and restore test passing
