### 2026-04-09 6eada78
fix: remove non-ASCII emoji from workflow for better locale compatibility
