### 2026-04-09 feac8fa
chore: finalize v0.2.3 with pkgcheck-action and standardized news
