+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: HydraR
 Type: Package
 Title: Stateful Agentic Orchestration for Scientific Reproducibility
-Version: 0.2.3
+Version: 0.2.4
 Authors@R: c(person(given = c("Chi", "Nam", "Ignatius"), family =
         "Pang", role = c("aut", "cre"), email =
         "ignatius.pang@mq.edu.au", comment = c(ORCID =
