### 2026-04-09 5cd2365
release: v0.2.3 final - aligning version and metadata
