+++ news/index.md
@@ -11,9 +11,11 @@
   - Replaced bitwise control characters (`\001`, `\002`) in
     `mermaid_parser.R` with safe string delimiters (`@@@`, `###`) to
     prevent signal interruptions in static analysis parsers.
-- **Workflow & Automation**: Updated the `pkgdown` GitHub workflow to
-  ensure unit tests are executed as a prerequisite for documentation
-  deployment.
+- **Workflow & Automation**:
+  - Updated the `pkgdown` GitHub workflow to ensure unit tests are
+    executed as a prerequisite for documentation deployment.
+  - Integrated the official `ropensci/pkgcheck-action` workflow to
+    automate rOpenSci compliance checks on every push.
 
 ## HydraR 0.2.2 (Documentation Overhaul & Release Polish)
 
