### 2026-04-10 88f1504
docs: regenerate pkgdown site for v0.2.3
