+++ docs/articles/data_science.md
@@ -20,6 +20,7 @@ re-optimization.
 ## Setup
 
 ``` r
+
 library(HydraR)
 ```
 
@@ -30,6 +31,7 @@ configuration, LLM prompts, agent roles, and deterministic logic—in a
 central registry.
 
 ``` r
+
 ds_logic_registry <- list(
   # 0. Initial Configuration
   initial_state = list(
@@ -89,6 +91,7 @@ We use a factory function to dynamically create nodes based on their
 type and parameters defined in the Mermaid graph.
 
 ``` r
+
 ds_node_factory <- function(id, label, params) {
   # Driver resolution from Mermaid params
   driver_obj <- if (!is.null(params$driver) && params$driver == "gemini") GeminiCLIDriver$new() else NULL
@@ -120,6 +123,7 @@ string serves as the single source of truth for both structure and node
 metadata.
 
 ``` r
+
 mermaid_graph <- "
 graph TD
   DataCleaner[Data Preprocessor] --> ModelTrainer
@@ -147,10 +151,12 @@ compiled_dag <- dag$compile()
 ## Visualizing the Workflow
 
 ``` r
+
 cat("```mermaid\n")
 ```
 
 ``` mermaid
+
 ``` r
 cat(compiled_dag$plot(type = "mermaid"))
 ```
@@ -178,8 +184,10 @@ graph TD
 ```
 
 ``` r
+
 cat("\n```\n")
 ```
+
 
     ## Running the Scenario
 
