### 2026-04-11 #766 @emilyriederer
Hi @IgnatiusPang ! Thanks for your patience. I am the rotating EiC and reviewing this package for fit.

A few questions for you:

- I broadly get the value of preserving state and auditability, but to help the editorial team more fully consider fit and categories, could you please describe a more concrete example use case? When would a researcher decide this was the right tool for their project?
- I note this is compatible with "Gemini, Claude, Copilot". Is the tool limited to proprietary models or is it suited to work with local and/or open source models as well?
