### 2026-04-14 #766 @IgnatiusPang
Thank you for taking the time to review HydraR as a rotating EiC and for the useful questions! I'd be happy to clarify.

* I broadly get the value of preserving state and auditability, but to help the editorial team more fully consider fit and categories, could you please describe a more concrete example use case? When would a researcher decide this was the right tool for their project?

HydraR is designed for scenarios where AI agents need to actively interact with the filesystem and execute complex toolchains to solve a problem, but where the "black-box" nature of typical AI wrappers would violate scientific reproducibility standards. While the tool is designed for a general scientific user base, my background is in bioinformatics, so I will stick to an example that I'm more familiar with. A concrete example is prototyping an integrative multi-omics data analysis pipeline. The integration of multi-omics data, including transcriptomics, proteomics, metabolomics, and lipidomics datasets, requires chaining disparate tools with varying data definitions. For example, using [edgeR]( https://bioconductor.org/packages/release/bioc/html/edgeR.html) for transcript expression, [limpa]( https://bioconductor.org/packages/release/bioc/html/limpa.html) for proteomics, [MetaboAnalystR](https://github.com/xia-lab/MetaboAnalystR) for metabolomics and lipidomics, visualising the volcano plots interactively using [Glimma](https://bioconductor.org/packages/release/bioc/html/Glimma.html), integrating the data using [MOFA2]( https://bioconductor.org/packages/release/bioc/html/MOFA2.html) or [mixOmics]( https://bioconductor.org/packages/release/bioc/html/mixOmics.html), and using StringDB and Reactome REST APIs for pathways and networks analysis. It is challenging for a researcher to learn to work with these tools and to choose parameters that adhere to best practices, and this process is time-consuming. These disparate steps could be too complicated for a single agent to perform in a single session. HydraR solves this by allowing an orchestrated team of AI agents to develop code that joins these steps autonomously and also to define templates for using these tools in a fully unified pipeline.

One unique strength of HydraR's state-preserving system is that its registry design allows the user to hot-swap the agent within the logical flow. Since all nodes are recorded in a “registry”, users can use swap_driver() (e.g. llm_node$swap_driver(“gpt4_api”)) to dynamically swap which LLM method to use based on logical tests before running a node. For example, an upstream node detected that the data is more complex than expected and can hot-swap to a more sophisticated model in one or more downstream nodes. Another example is that a failover node can detect that the node directly upstream used a weaker model, which led to an error, then hot-swap to a more sophisticated model, loop back, and re-run that step. This way, it's a self-healing property that users can manually implement in the network using the state recording registry.

Due to the directed acyclic graph design, git-based state isolation, and the ability to enable structured logging in the duckDB database, there is a very high level of traceability throughout the entire process. This level of traceability makes the tool suitable for applications that require high auditability and compliance checking. If the study's findings are ever questioned, an auditor can trace the exact lineage of an AI's decision-making process by reviewing the logs and git's commit history, tracing the chain of evidence from raw input to final output.

Beyond research, HydraR serves as a platform for innovation. By providing these orchestration primitives (Dynamic Nodes, DAGs, Registries) natively in R, we enable developers to experiment with new multi-agent architectures that move beyond simple sequential scripts and into adaptive, autonomous systems.

* I note this is compatible with "Gemini, Claude, Copilot". Is the tool limited to proprietary models, or is it suited to work with local and/or open source models as well?

The HydraR tool enables users to use any model, both open-sourced and commercial. I have already enabled the nodes to work with Ollama via our OllamaDriver, and I have tested that it works with a small model, but I am sure the mechanics would allow it to work with a larger one. The AgentDriver R6 architecture is intended to be extensible, so that it is model-agnostic and anyone can add any type of drivers in the future, including CLI-based, API-based, and locally hosted models. An example of this is on the [Creating Custom Agent Driver page](https://github.com/APAF-bioinformatics/HydraR/blob/main/vignettes/extending_hydrar.md).
