### 2026-04-17 #766 @mpadge
@IgnatiusPang We've been doing some updates to the checking system, and noticed a bug in processing your package that failed to identify some issues. That's not deployed yet, but when we get that far, the report will fail because of several uses of global assignment operators (`<<-`) throughout your code. A lot of these seem to be in  R6 `self <<-` assignment, which don't seem necessary at all. Can you please try to remove all of these, or at least explain the necessity of any that you deem unavoidable? Thanks!
