### 2026-04-17 32961cd
release: v0.2.4 - APAF compliance refactor (remove <<-) and introduce AgentHasher & skill discovery
