### 2026-04-17 5d25d63
fix: restore AgentDAG signature and fix skill discovery test syntax
