### 2026-04-17 72952c5
chore: isolate AgentHasher and skill discovery to dev branch; v0.2.4 focused on compliance
