+++ tests/testthat/test-blog_writer.R
@@ -67,15 +67,16 @@ test_that("Blog Writer DAG loop works", {
   dag$add_edge("Drafter", "Editor")
 
   # Inject loop success on second iteration
-  loop_count <- 0
+  env <- new.env(parent = emptyenv())
+  env$loop_count <- 0
   dag$add_conditional_edge(
     from = "Editor",
     test = function(out) {
-      loop_count <<- loop_count + 1
-      if (loop_count == 1) {
+      env$loop_count <- env$loop_count + 1
+      if (env$loop_count == 1) {
         driver$response <- "Approved"
       }
-      if (loop_count >= 2) {
+      if (env$loop_count >= 2) {
         return(TRUE)
       }
       return(FALSE)
@@ -91,7 +92,7 @@ test_that("Blog Writer DAG loop works", {
     initial_state = list(blog_topic = "R Agents"),
     max_steps = 10
   ))
-  expect_equal(loop_count, 2)
+  expect_equal(env$loop_count, 2)
   expect_equal(result$state$get("Editor"), "Approved")
 })
 
