+++ news/index.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## HydraR 0.2.4 (APAF Compliance & Idempotency)
+## HydraR 0.2.4 (APAF Compliance Refactor)
 
 - **APAF Compliance Refactor**:
   - Absolute removal of all global assignment operators (`<<-`) across
@@ -9,15 +9,6 @@
     ([`purrr::reduce`](https://purrr.tidyverse.org/reference/reduce.html)).
   - Standardized state mutation within R6 methods to use direct field
     assignment (`self$x <- val`).
-- **Native Idempotency Engine**:
-  - Introduced `AgentHasher` (R/hasher.R), a DuckDB-backed engine for
-    node-level execution hashing.
-  - Enables “targets-like” behavior where nodes can skip execution if
-    input state and logic remain unchanged.
-- **Automated Skill Discovery**:
-  - Implemented `discover_package_skills()` (R/skill_discovery.R) for
-    automatic discovery and registration of external package skills via
-    `hydrar/manifest.yaml`.
 - **Driver & Execution Enhancements**:
   - Refactored `AgentLLMNode` to use
     [`purrr::reduce`](https://purrr.tidyverse.org/reference/reduce.html)
