### 2026-04-17 #766 @IgnatiusPang
Thanks @mpadge, I think I had refactored all instances of self <<- assignment in the code. I have updated the code to a new version v0.2.4 and will recheck the package using the bot.
