### 2026-04-17 587bc59
docs: finalize v0.2.4 release notes for APAF compliance
