+++ NEWS.md
@@ -3,10 +3,6 @@
 - **APAF Compliance Refactor**:
     - Absolute removal of all global assignment operators (`<<-`) across the package (R scripts, tests, and vignettes) in favor of local environments and functional programming patterns (`purrr::reduce`).
     - Standardized state mutation within R6 methods to use direct field assignment (`self$x <- val`).
-- **Driver & Execution Enhancements**:
-    - Refactored `AgentLLMNode` to use `purrr::reduce` for robust system prompt construction.
-    - Improved environment isolation for Python and Bash nodes.
-    - Enhanced error handling for `reticulate`-based Python execution.
 - **Unit Testing Overhaul**:
     - Updated the entire test suite to be `<<-` clean.
     - Improved mock driver implementations for better `httr2` and CLI simulation.
