### 2026-04-17 #766 @mpadge
@IgnatiusPang Sorry if i wasn't clear enough - those changes to our checking system have not yet been deployed. You checks will likely remain as they were previously. New deployment should arrive Tues 21st Apr. You can still update `pkgcheck` locally to see results of latest check updates.
