### 2026-04-23 23689aa
docs: add website URL to DESCRIPTION and _pkgdown.yml to resolve pkgdown check
