+++ docs/reference/AgentMapNode.md
@@ -8,7 +8,7 @@ An \`AgentMapNode\` object.
 
 ## Super class
 
-[`HydraR::AgentNode`](https://github.com/APAF-bioinformatics/HydraR/reference/AgentNode.md)
+[`HydraR::AgentNode`](https://APAF-bioinformatics.github.io/HydraR/reference/AgentNode.md)
 -\> `AgentMapNode`
 
 ## Public fields
