+++ reference/OllamaDriver.md
@@ -15,7 +15,7 @@ specify the binary path via `options(HydraR.ollama_path = "...")`.
 
 ## Super class
 
-[`HydraR::AgentDriver`](https://github.com/APAF-bioinformatics/HydraR/reference/AgentDriver.md)
+[`HydraR::AgentDriver`](https://APAF-bioinformatics.github.io/HydraR/reference/AgentDriver.md)
 -\> `OllamaDriver`
 
 ## Public fields
@@ -38,11 +38,11 @@ specify the binary path via `options(HydraR.ollama_path = "...")`.
 
 Inherited methods
 
-- [`HydraR::AgentDriver$exec_in_dir()`](https://github.com/APAF-bioinformatics/HydraR/reference/AgentDriver.html#method-exec_in_dir)
-- [`HydraR::AgentDriver$filter_llm_noise()`](https://github.com/APAF-bioinformatics/HydraR/reference/AgentDriver.html#method-filter_llm_noise)
-- [`HydraR::AgentDriver$get_capabilities()`](https://github.com/APAF-bioinformatics/HydraR/reference/AgentDriver.html#method-get_capabilities)
-- [`HydraR::AgentDriver$validate_cli_opts()`](https://github.com/APAF-bioinformatics/HydraR/reference/AgentDriver.html#method-validate_cli_opts)
-- [`HydraR::AgentDriver$validate_no_injection()`](https://github.com/APAF-bioinformatics/HydraR/reference/AgentDriver.html#method-validate_no_injection)
+- [`HydraR::AgentDriver$exec_in_dir()`](https://APAF-bioinformatics.github.io/HydraR/reference/AgentDriver.html#method-exec_in_dir)
+- [`HydraR::AgentDriver$filter_llm_noise()`](https://APAF-bioinformatics.github.io/HydraR/reference/AgentDriver.html#method-filter_llm_noise)
+- [`HydraR::AgentDriver$get_capabilities()`](https://APAF-bioinformatics.github.io/HydraR/reference/AgentDriver.html#method-get_capabilities)
+- [`HydraR::AgentDriver$validate_cli_opts()`](https://APAF-bioinformatics.github.io/HydraR/reference/AgentDriver.html#method-validate_cli_opts)
+- [`HydraR::AgentDriver$validate_no_injection()`](https://APAF-bioinformatics.github.io/HydraR/reference/AgentDriver.html#method-validate_no_injection)
 
 ------------------------------------------------------------------------
 
