### 2026-04-23 ceb3c63
docs: reorder URLs in DESCRIPTION and add URL to codemeta.json to fix website check
