+++ pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   sorting_benchmark: sorting_benchmark.html
   state_persistence: state_persistence.html
   targets_integration: targets_integration.html
-last_built: 2026-04-23T12:17Z
+last_built: 2026-04-23T14:40Z
 urls:
   reference: https://APAF-bioinformatics.github.io/HydraR/reference
   article: https://APAF-bioinformatics.github.io/HydraR/articles
