### 2026-04-23 #766 @IgnatiusPang
I ran the R open science package check above as per the instructions from [mpadge](https://github.com/mpadge), which led to the discovery of the error with my webpage. I have now fixed this error, rerun the package check in my own GitHub Actions, and everything has passed.
