### 2026-05-03 #766 @maurolepore
Thanks for working through the package-check items. The bot pkgcheck is currently failing with a service-side error we've been seeing on other submissions — it's not a problem with your package. I'll re-trigger in a few days; once it passes, the next step is to assign a handling editor.

For the record, in the statistical properties of the last pkgcheck run I see the package has a relatively large R/ directory and number of vignettes (around the 95% percentile).

Thanks for sharing your work with the rOpenSci community.
